requirements.txt
django
design-django-theme
cryptojwt
requests
python-magic
phonenumbers
zeep
pypdf

weasyprint
# sudo apt install libpango-1.0-0 libcairo2 libgdk-pixbuf2.0-0 libffi-dev libssl-dev
